requirements.txt
Django>=4.2.0
Pillow>=10.0.0
django-ckeditor>=6.7.0
django-cleanup>=7.0.0
django-extensions>=3.2.3
python-decouple>=3.8

